add proxy-components for card
