+++ src/components/features/card/card-favorites.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { CardProps } from '../types';
+import { CardType } from '../../../const';
+import Card from './card';
+
+function CardCities(props: CardProps): React.ReactElement {
+  const { className = '', cardData } = props;
+
+  return (
+    <Card
+      className={`favorites__card ${className}`}
+      cardData={cardData}
+      cardType={CardType.FAVORITES}
+    />
+  );
+}
+
+export default CardCities;
