+++ src/components/features/types.ts
@@ -1,8 +1,11 @@
+import React from 'react';
 import { Hotel } from '../../data-type';
 
 export interface CardProps {
   cardData: Hotel,
-  cardType?: string
+  cardType?: string,
+  className?: string,
+  children?: React.ReactElement,
 }
 
 export interface OffersListProps {
